bot/requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0
